The 64-bit MSC warning hunt stops for today.
svn:sync-xref-src-repo-rev: r39131

+++ src/VBox/VMM/PATM/CSAM.cpp
@@ -1,4 +1,4 @@
-/* $Id: CSAM.cpp 13827 2008-11-05 01:31:18Z knut.osmundsen@oracle.com $ */
+/* $Id: CSAM.cpp 14075 2008-11-11 00:14:05Z knut.osmundsen@oracle.com $ */
 /** @file
  * CSAM - Guest OS Code Scanning and Analysis Manager
  */
@@ -1915,7 +1915,7 @@ static int csamRemovePageRecord(PVM pVM, RTRCPTR GCPtr)
  */
 static DECLCALLBACK(void) CSAMDelayedWriteHandler(PVM pVM, RTRCPTR GCPtr, size_t cbBuf)
 {
-    int rc = PATMR3PatchWrite(pVM, GCPtr, cbBuf);
+    int rc = PATMR3PatchWrite(pVM, GCPtr, (uint32_t)cbBuf);
     AssertRC(rc);
 }
 
@@ -1944,7 +1944,7 @@ static DECLCALLBACK(int) CSAMCodePageWriteHandler(PVM pVM, RTGCPTR GCPtr, void *
 
     if (VM_IS_EMT(pVM))
     {
-        rc = PATMR3PatchWrite(pVM, GCPtr, cbBuf);
+        rc = PATMR3PatchWrite(pVM, GCPtr, (uint32_t)cbBuf);
     }
     else
     {
